Refactor code,centralized get details from movies and tv shows
