+++ src/core/films/section/tv.component.js
@@ -3,18 +3,19 @@ import { useHistory } from "react-router-dom";
 
 //own
 import useInitialiceTv from "../../main/use-initialice-tv";
-import { detailActions } from "../../../store/collections/movies/details/detail-slice";
+import { detailActions } from "../../../store/collections/details/detail-slice";
 import Spinner from "../../../components/spinner";
 import Tv from "../tv";
 
 const TvSection = (props) => {
   useInitialiceTv();
+
   const dispatch = useDispatch();
   const history = useHistory();
   const { tvPopular, isLoadingTv, showModal } = useSelector((state) => ({
     tvPopular: state.tvCollection.tvPopular,
     isLoadingTv: state.tvCollection.isLoadingTv,
-    showModal: state.videosCollection.showModal,
+    showModal: state.detailsCollection.showModal,
   }));
 
   const handleModalTv = (category, id) => {
