+++ src/core/main/use-initialice-videos.js
@@ -2,8 +2,8 @@ import { useEffect } from "react";
 import { useDispatch } from "react-redux";
 import { useParams } from "react-router-dom";
 
-import { getDetailMovies } from "../../store/collections/movies/details/get-detail";
-import { detailActions } from "../../store/collections/movies/details/detail-slice";
+import { getDetailFilm } from "../../store/collections/details/get-detail";
+import { detailActions } from "../../store/collections/details/detail-slice";
 
 export default function useInitialiceVideos() {
   const dispatch = useDispatch();
@@ -12,6 +12,6 @@ export default function useInitialiceVideos() {
 
   useEffect(() => {
     dispatch(detailActions.setDetails([]));
-    dispatch(getDetailMovies(collectionId, typeCollection));
+    dispatch(getDetailFilm(collectionId, typeCollection));
   }, [dispatch, collectionId, typeCollection]);
 }
