+++ src/store/collections/details/detail-slice.js
