+++ src/store/index.js
@@ -1,6 +1,5 @@
 import { configureStore } from "@reduxjs/toolkit";
 import moviesSlice from "./collections/movies/movies-slice";
-import videosTvSlice from "./collections/tv/details/detail-tv-slice";
 import tvSlice from "./collections/tv/tv-slice";
 import genresSlice from "./genres/genres-slice";
 import detailSlice from "./collections/movies/details/detail-slice";
@@ -13,7 +12,6 @@ const store = configureStore({
     genresCollection: genresSlice.reducer,
     detailsCollection: detailSlice.reducer,
     tvCollection: tvSlice.reducer,
-    videosTvCollection: videosTvSlice.reducer,
     searchResults: searchSlice.reducer,
     topRatedCollection: topRatedSlice.reducer,
   },
